+++ packages/adapters/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamba/adapters
 
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamba/core@0.18.0
+
 ## 0.17.0
 
 ### Patch Changes